Update bonus-fastify-joi.md
You could use either the version I committed — see below:

"If you are an hapi developer, you may know the joi library. It is a powerful validation library that allows you to define programmatic schemas for your data and validate them with ease."

Or 

"If you are an hapi developer, you may know the joi library. It is a powerful validation library that allows you to programmatically define schemas for your data and validate them with ease."