Update bonus-on-listen.md
Excellent work 🌟.

I made a few tiny changes. Almost all were changing the capitalisation of the subheadings.

The main change for you to be aware of is the one below in quotation marks:

While we won't delve into Request/Reply hooks in this article, you can find comprehensive information in the Fastify documentation "or in chapter 4 of the detailed Fastify book 📙!"

There was nothing wrong with what you wrote, I just made the above change because it reads a little more naturally.